fix: move lifecycle handling into renderer component
